Allow the code to run on followers
From the documentaiton txid_current():
get current transaction ID, assigning a new one if the current
transaction does not have one.

This prevents us to use txid_current() on the follower.
Getting txid_current_snapshot allows us to get the xmax value from the
current snapshot, and the code will run on both the master and the
follower.