Support Elasticsearch v8 (and v7) & gem v1 release
We are still on v0 so going for a breaking change since we no longer care
about Elasticsearch < v7 and Ruby < 2.7.

But also working towards a v1 release since the gem has been somewhat
neglected, by doing some extra changes to remove hacks in other
services, e.g. deals service.

Added
- Support for Ruby 3 (and keep support for 2.7).
- Support for Elasticsearch v8 (and keep support for v7).
- Support setting a logger in `Config`.
- Support refresh on `IndexManager#populate_index`.
- Support Proc in `Config#data_source` so it can be lazily evaluated.

Removed
- Drop support for Ruby 2.6. (Previous PR.)
- Drop support for Elasticsearch v5 and v6.

The README has also been updated to reflect that the gem only supports
Elasticsearch and ActiveRecord with PG since we rely on txid which is PG
specific and we don't have a way around it.